+++ benchmarks/src/main/java/org/questdb/LineUDPSenderMain.java
@@ -24,6 +24,7 @@
 
 package org.questdb;
 
+import io.questdb.cutlass.line.AbstractLineSender;
 import io.questdb.cutlass.line.LineUdpSender;
 import io.questdb.network.Net;
 import io.questdb.std.Os;
