+++ core/src/test/java/io/questdb/cutlass/line/tcp/LineTcpReceiverTest.java
@@ -31,8 +31,8 @@ import io.questdb.cairo.security.AllowAllCairoSecurityContext;
 import io.questdb.cairo.sql.ReaderOutOfDateException;
 import io.questdb.cairo.sql.RecordCursor;
 import io.questdb.cairo.sql.RecordCursorFactory;
+import io.questdb.cutlass.line.AbstractLineSender;
 import io.questdb.cutlass.line.AuthenticatedLineTcpSender;
-import io.questdb.cutlass.line.LineUdpSender;
 import io.questdb.cutlass.line.LineTcpSender;
 import io.questdb.griffin.CompiledQuery;
 import io.questdb.griffin.SqlCompiler;
@@ -63,7 +63,6 @@ import org.junit.Test;
 import java.net.URL;
 import java.nio.charset.StandardCharsets;
 import java.security.PrivateKey;
-import java.sql.Time;
 import java.util.concurrent.TimeUnit;
 import java.util.concurrent.locks.LockSupport;
 
@@ -192,7 +191,7 @@ public class LineTcpReceiverTest extends AbstractCairoTest {
                         "weather windspeed=4.0 631170000000000000\n";
 
         runInContext((receiver) -> {
-            send(receiver, lineData, "weather", WAIT_ENGINE_TABLE_RELEASE, false);
+            sendLinger(receiver, lineData, "weather");
 
             String expected =
                     "windspeed\ttimestamp\ttimetocycle\n" +
@@ -212,7 +211,7 @@ public class LineTcpReceiverTest extends AbstractCairoTest {
                         "weather windspeed=4.0 631170000000000000\n";
 
         runInContext((receiver) -> {
-            send(receiver, lineData, "weather", WAIT_ENGINE_TABLE_RELEASE, false);
+            sendLinger(receiver, lineData, "weather");
 
             String expected =
                     "windspeed\ttimestamp\ttimetocycle\n" +
@@ -233,7 +232,7 @@ public class LineTcpReceiverTest extends AbstractCairoTest {
                         "weather dir=\"SSW\",windspeed=4.0 631170000000000000\n";
 
         runInContext((receiver) -> {
-            send(receiver, lineData, "weather", WAIT_ENGINE_TABLE_RELEASE, false);
+            sendLinger(receiver, lineData, "weather");
 
             String expected =
                     "dir\twindspeed\ttimestamp\ttimetocycle\n" +
@@ -369,7 +368,7 @@ public class LineTcpReceiverTest extends AbstractCairoTest {
                     "plug,room=6A watts=\"3188\" 1631817599910\n" +
                     "plug,room=6A watts=\"3180\" 1631817902842\n" +
                     "plug,label=Power,room=6A watts=\"475\" 1631817478737\n";
-            send(receiver, lineData, "plug", WAIT_ENGINE_TABLE_RELEASE, false);
+            sendLinger(receiver, lineData, "plug");
 
             String expected = "room\twatts\ttimestamp\tlabel\n" +
                     "6A\t3195\t1970-01-01T00:27:11.817296Z\t\n" +
@@ -387,7 +386,7 @@ public class LineTcpReceiverTest extends AbstractCairoTest {
         runInContext((receiver) -> {
             String lineData = "plug,room=6A watts=\"1\" 2631819999000\n" +
                     "plug,label=Power,room=6B watts=\"22\" 1631817902842\n";
-            send(receiver, lineData, "plug", WAIT_ENGINE_TABLE_RELEASE, false);
+            sendLinger(receiver, lineData, "plug");
 
             String expected = "room\twatts\ttimestamp\tlabel\n" +
                     "6B\t22\t1970-01-01T00:27:11.817902Z\tPower\n" +
@@ -403,7 +402,7 @@ public class LineTcpReceiverTest extends AbstractCairoTest {
             String lineData = "plug,room=6A watts=\"1\" 2631819999000\n" +
                     "plug,label=Power,room=6B watts=\"22\" 1631817902842\n" +
                     "plug,label=Line,room=6C watts=\"333\" 1531817902842\n";
-            send(receiver, lineData, "plug", WAIT_ENGINE_TABLE_RELEASE, false);
+            sendLinger(receiver, lineData, "plug");
 
             String expected = "room\twatts\ttimestamp\tlabel\n" +
                     "6C\t333\t1970-01-01T00:25:31.817902Z\tLine\n" +
@@ -447,7 +446,7 @@ public class LineTcpReceiverTest extends AbstractCairoTest {
                                     "weather windspeed=2.0 631152000000000000\n" +
                                     "weather timetocycle=0.0,windspeed=3.0 631160000000000000\n" +
                                     "weather windspeed=4.0 631170000000000000\n";
-                    send(receiver, lineData, "weather", WAIT_ENGINE_TABLE_RELEASE, false);
+                    sendLinger(receiver, lineData, "weather");
                 });
 
                 try (RecordCursor cursor = cursorFactory.getCursor(sqlExecutionContext)) {
@@ -484,7 +483,7 @@ public class LineTcpReceiverTest extends AbstractCairoTest {
         String lineData = "table_a,MessageType=B,SequenceNumber=1 Length=92i,test=1.5 1465839830100400000\n";
 
         runInContext((receiver) -> {
-            send(receiver, lineData, "table_a", WAIT_ENGINE_TABLE_RELEASE, false);
+            sendLinger(receiver, lineData, "table_a");
 
             String expected = "ReceiveTime\tSequenceNumber\tMessageType\tLength\ttest\n" +
                     "2016-06-13T17:43:50.100400Z\t1\tB\t92\t1.5\n";
@@ -516,7 +515,7 @@ public class LineTcpReceiverTest extends AbstractCairoTest {
                                 "up in=2.0 631152000000000000\n" +
                                 "up in=3.0 631160000000000000\n" +
                                 "up in=4.0 631170000000000000\n";
-                send(receiver, lineData, "up", WAIT_ENGINE_TABLE_RELEASE, false);
+                sendLinger(receiver, lineData, "up");
             });
 
             TestUtils.assertSql(compiler, sqlExecutionContext, "up", sink, expected);
@@ -530,7 +529,7 @@ public class LineTcpReceiverTest extends AbstractCairoTest {
                 "tag_n_7=7,tag_n_8=8,tag_n_9=9,tag_n_10=10,tag_n_11=11,tag_n_12=12,tag_n_13=13," +
                 "tag_n_14=14,tag_n_15=15,tag_n_16=16,tag_n_17=17 value=42.4 1619509249714000000\n";
         runInContext((receiver) -> {
-            send(receiver, lineData, "tableCRASH", WAIT_ENGINE_TABLE_RELEASE, false);
+            sendLinger(receiver, lineData, "tableCRASH");
 
             String expected = "tag_n_1\ttag_n_2\ttag_n_3\ttag_n_4\ttag_n_5\ttag_n_6\ttag_n_7\ttag_n_8\ttag_n_9\ttag_n_10\ttag_n_11\ttag_n_12\ttag_n_13\ttag_n_14\ttag_n_15\ttag_n_16\ttag_n_17\tvalue\ttimestamp\n" +
                     "1\t2\t3\t4\t5\t6\t7\t8\t9\t10\t11\t12\t13\t14\t15\t16\t17\t42.400000000000006\t2021-04-27T07:40:49.714000Z\n";
@@ -549,7 +548,7 @@ public class LineTcpReceiverTest extends AbstractCairoTest {
                 + "tab ts_nsec=5555555555555555555i,raw_msg=\"_________________________________________________________________________________________________________ ____________\" 1619509249714000000\n"
                 + "tab ts_nsec=6666666666666666666i,raw_msg=\"_________________________________________________________________________________________________________ ____________\" 1619509249714000000\n";
         runInContext((receiver) -> {
-            send(receiver, lineData, "tab", WAIT_ENGINE_TABLE_RELEASE, false);
+            sendLinger(receiver, lineData, "tab");
 
             String expected = "ts_nsec\traw_msg\ttimestamp\n" +
                     "1111111111111111111\t_________________________________________________________________________________________________________ ____________\t2021-04-27T07:40:49.714000Z\n" +
@@ -565,8 +564,8 @@ public class LineTcpReceiverTest extends AbstractCairoTest {
     @Test
     public void testFieldWithUnquotedString() throws Exception {
         runInContext((receiver) -> {
-            send(receiver,  "tab raw_msg=____ 1619509249714000000\n", "tab", WAIT_ENGINE_TABLE_RELEASE, false);
-            send(receiver,  "tab raw_msg=__\"_ 1619509249714000000\n", "tab", WAIT_ENGINE_TABLE_RELEASE, false);
+            sendLinger(receiver,  "tab raw_msg=____ 1619509249714000000\n", "tab");
+            sendLinger(receiver,  "tab raw_msg=__\"_ 1619509249714000000\n", "tab");
 
             String expected = "raw_msg\ttimestamp\n" +
                     "____\t2021-04-27T07:40:49.714000Z\n" +
@@ -589,10 +588,10 @@ public class LineTcpReceiverTest extends AbstractCairoTest {
 
         runInContext((receiver) -> {
             String lineData = "लаблअца поле=\"значение\" 1619509249714000000\n";
-            send(receiver, lineData, "लаблअца", WAIT_ENGINE_TABLE_RELEASE, false);
+            sendLinger(receiver, lineData, "लаблअца");
 
             String lineData2 = "लаблअца,символ=значение2 поле=\"значение3\" 1619509249714000000\n";
-            send(receiver, lineData2, "लаблअца", WAIT_ENGINE_TABLE_RELEASE, false);
+            sendLinger(receiver, lineData2, "लаблअца");
 
             assertTable("символ\tполе\tвремя\n" +
                     "\tзначение\t2021-04-27T07:40:49.714000Z\n" +
@@ -604,10 +603,10 @@ public class LineTcpReceiverTest extends AbstractCairoTest {
     public void testUnicodeTableNameExistingTable() throws Exception {
         runInContext((receiver) -> {
             String lineData = "लаблअца поле=значение 1619509249714000000\n";
-            send(receiver, lineData, "लаблअца", WAIT_ENGINE_TABLE_RELEASE, false);
+            sendLinger(receiver, lineData, "लаблअца");
 
             String lineData2 = "लаблअца,символ=значение2  1619509249714000000\n";
-            send(receiver, lineData2, "लаблअца", WAIT_ENGINE_TABLE_RELEASE, false);
+            sendLinger(receiver, lineData2, "लаблअца");
 
             String expected = "поле\ttimestamp\tсимвол\n" +
                     "значение\t2021-04-27T07:40:49.714000Z\t\n" +
@@ -775,7 +774,7 @@ public class LineTcpReceiverTest extends AbstractCairoTest {
             };
 
             String lineData = "table_a,MessageType=B,SequenceNumber=1 Length=92i,test=1.5 1465839830100400000\n";
-            send(receiver, lineData, "table_a", WAIT_ENGINE_TABLE_RELEASE, false);
+            sendLinger(receiver, lineData, "table_a");
 
             String expected = "ReceiveTime\tSequenceNumber\tMessageType\tLength\n";
             assertTable(expected, "table_a");
@@ -938,15 +937,11 @@ public class LineTcpReceiverTest extends AbstractCairoTest {
     }
 
     private void send(LineTcpReceiver receiver, String lineData, String tableName, int wait) {
-        send(receiver, tableName, wait, () -> {
-            sendToSocket(lineData, true);
-        });
+        send(receiver, tableName, wait, () -> sendToSocket(lineData, true));
     }
 
-    private void send(LineTcpReceiver receiver, String lineData, String tableName, int wait, boolean nolinger) {
-        send(receiver, tableName, wait, () -> {
-            sendToSocket(lineData, nolinger);
-        });
+    private void sendLinger(LineTcpReceiver receiver, String lineData, String tableName) {
+        send(receiver, tableName, LineTcpReceiverTest.WAIT_ENGINE_TABLE_RELEASE, () -> sendToSocket(lineData, false));
     }
 
     public static final int WAIT_NO_WAIT = 0;
@@ -1070,7 +1065,7 @@ public class LineTcpReceiverTest extends AbstractCairoTest {
                 sharedWorkerPool.start(LOG);
 
                 try {
-                    final LineUdpSender[] senders = new LineUdpSender[tables.size()];
+                    final AbstractLineSender[] senders = new AbstractLineSender[tables.size()];
                     for (int n = 0; n < senders.length; n++) {
                         if (null != authKeyId) {
                             AuthenticatedLineTcpSender sender = new AuthenticatedLineTcpSender(
@@ -1094,7 +1089,7 @@ public class LineTcpReceiverTest extends AbstractCairoTest {
                     StringSink tsSink = new StringSink();
                     for (int nRow = 0; nRow < nRows; nRow++) {
                         int nTable = nRow < tables.size() ? nRow : rand.nextInt(tables.size());
-                        LineUdpSender sender = senders[nTable];
+                        AbstractLineSender sender = senders[nTable];
                         StringBuilder sb = expectedSbs[nTable];
                         CharSequence tableName = tables.get(nTable);
                         sender.metric(tableName);
@@ -1120,7 +1115,7 @@ public class LineTcpReceiverTest extends AbstractCairoTest {
                     }
 
                     for (int n = 0; n < senders.length; n++) {
-                        LineUdpSender sender = senders[n];
+                        AbstractLineSender sender = senders[n];
                         sender.close();
                     }
